refactor: post 컴포넌트 contents(로직중심), components(ui 중심)로 분리
