+++ src/app/post/[slug]/loading.tsx
@@ -1,5 +1,5 @@
-import PostPageSkeleton from '@/features/post/public/postDetail/components/PostPageSkeleton';
+import PostDetailSkeleton from '@/features/post/public/contents/postDetail/PostDetailSkeleton';
 
 export default function Loading() {
-  return <PostPageSkeleton />;
+  return <PostDetailSkeleton />;
 }
