+++ src/features/post/public/components/postDetail/PostInformation.tsx
