+++ src/features/post/public/store/PostSearchStateStore.tsx
